Consolidate Copilot App environment scanning

Copilot-Session: dfca4c57-330a-406c-b8d2-eada8bddec9a
